refactor: recoil selector 이용해서 filteredRestaurant 수정
