+++ src/components/Main/CategoryFilter.jsx
@@ -1,5 +1,7 @@
 import { categories } from "../../data/data.js";
 import styled from "styled-components";
+import { useRecoilState } from "recoil";
+import { categoryState } from "../../store/AppAtom.jsx";
 
 const RestaurantFilterContainer = styled.section`
   display: flex;
@@ -24,7 +26,12 @@ const RestaurantFilter = styled.select`
   padding: 8px;
 `;
 
-function CategoryFilter({ category, onChangeCategory }) {
+function CategoryFilter() {
+  const [category, setCategory] = useRecoilState(categoryState);
+  const handleCategoryChange = (newCategory) => {
+    setCategory(newCategory);
+  };
+
   return (
     <RestaurantFilterContainer>
       <RestaurantFilter
@@ -32,7 +39,7 @@ function CategoryFilter({ category, onChangeCategory }) {
         id="category-filter"
         aria-label="음식점 카테고리 필터"
         value={category}
-        onChange={(event) => onChangeCategory(event.target.value)}
+        onChange={(event) => handleCategoryChange(event.target.value)}
       >
         {categories.map((cat) => (
           <option key={cat.id} value={cat.value}>
