+++ src/store/AppAtom.jsx
@@ -1,4 +1,4 @@
-import { atom } from "recoil";
+import { atom, selector } from "recoil";
 
 export const modalTypeState = atom({
   key: "modalTypeState",
@@ -13,4 +13,19 @@ export const clickedRestaurantInfoState = atom({
 export const restaurantsState = atom({
   key: "restaurantsState",
   default: [],
+});
+
+export const categoryState = atom({
+  key: "categoryState",
+  default: "전체",
+});
+
+export const filteredRestaurantsState = selector({
+  key: "filterdRestaurantsState",
+  get: ({ get }) => {
+    const all = get(restaurantsState);
+    const category = get(categoryState);
+    if (category === "전체") return all;
+    return all.filter((r) => r.category === category);
+  },
 });
